requirements.txt
amqp==5.0.6
asgiref==3.3.1
autopep8==1.5.6
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.7
django-crispy-forms==1.11.2
django-debug-toolbar==3.2
django-environ==0.4.5
django-lifecycle==0.9.1
django-redis==4.12.1
django-silk==4.1.0
django-split-settings==1.0.1
djangorestframework==3.12.2
gprof2dot==2021.2.21
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
kombu==5.0.2
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
Pillow==8.1.2
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pyflakes==2.2.0
Pygments==2.8.1
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
urlman==1.4.0
vine==5.0.0
wcwidth==0.2.5
